test: Seclude `resolveTestScenariosConfig`
